Refactor: add enums.py
